+++ example/lib/page/video_page_view_page.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_video_caching/ext/log_ext.dart';
 import 'package:flutter_video_caching/flutter_video_caching.dart';
 import 'package:video_player/video_player.dart';
 
@@ -56,6 +57,7 @@ class _VideoPageViewPageState extends State<VideoPageViewPage> {
   void dispose() {
     timer?.cancel();
     pageController.dispose();
+    VideoProxy.downloadManager.removeAllTask();
     super.dispose();
   }
 }
@@ -105,7 +107,7 @@ class _VideoPlayerWidgetState extends State<VideoPlayerWidget> {
 
   void playListener() {
     if (playControl.value.hasError) {
-      print("errorDescription: ${playControl.value.errorDescription}");
+      logE("errorDescription: ${playControl.value.errorDescription}");
       if (playControl.value.errorDescription!.contains("Source error")) {
         Uri uri = Uri.parse(widget.url);
         initPlayControl(uri);