+++ example/lib/page/video_play_page.dart
@@ -1,4 +1,7 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
+import 'package:flutter_video_caching/ext/log_ext.dart';
 import 'package:flutter_video_caching/flutter_video_caching.dart';
 import 'package:video_player/video_player.dart';
 
@@ -11,6 +14,7 @@ class VideoPlayPage extends StatefulWidget {
 
 class _VideoPlayPageState extends State<VideoPlayPage> {
   late VideoPlayerController _controller;
+  Timer? timer;
 
   @override
   void initState() {
@@ -20,21 +24,25 @@ class _VideoPlayPageState extends State<VideoPlayPage> {
 
   void initController() {
     List<String> urls = [
-      // 'http://vjs.zencdn.net/v/oceans.mp4',
-      'https://vv.jisuzyv.com/play/DbDGZ8ka/index.m3u8',
+      'http://vjs.zencdn.net/v/oceans.mp4',
+      'https://customer-fzuyyy7va6ohx90h.cloudflarestream.com/861279ab37d84dbfbf3247322fbcfc63/manifest/video.m3u8',
       'https://test-streams.mux.dev/x36xhzz/url_6/193039199_mp4_h264_aac_hq_7.m3u8',
+      'https://vv.jisuzyv.com/play/DbDGZ8ka/index.m3u8',
     ];
     var url = urls[0];
     String localUri = url.toLocalUrl();
     String remoteUri = localUri.toOriginUrl();
-    print('localUri: $localUri');
-    print('remoteUri: $remoteUri');
+    logD('localUri: $localUri');
+    logD('remoteUri: $remoteUri');
     Uri uri = url.toLocalUri();
     // Uri uri = Uri.parse(remoteUri);
     _controller = VideoPlayerController.networkUrl(uri)
       ..setLooping(true)
       ..initialize().then((_) {
         setState(() {});
+        timer = Timer(const Duration(milliseconds: 200), () {
+          _controller.play();
+        });
       });
   }
 
@@ -91,6 +99,7 @@ class _VideoPlayPageState extends State<VideoPlayPage> {
 
   @override
   void dispose() {
+    timer?.cancel();
     _controller.dispose();
     VideoProxy.downloadManager.removeAllTask();
     super.dispose();