+++ lib/parser/url_parser_mp4.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
@@ -15,11 +16,6 @@ import 'url_parser.dart';
 
 /// MP4 URL parser
 class UrlParserMp4 implements UrlParser {
-  @override
-  bool match(Uri uri) {
-    return uri.path.toLowerCase().endsWith('.mp4');
-  }
-
   /// Get the cache data from memory or file.
   /// If there is no cache data, return null.
   @override
@@ -85,87 +81,207 @@ class UrlParserMp4 implements UrlParser {
     try {
       RegExp exp = RegExp(r'bytes=(\d+)-(\d*)');
       RegExpMatch? rangeMatch = exp.firstMatch(headers['range'] ?? '');
-      int rangeStart = int.parse(rangeMatch?.group(1) ?? '0');
-      String responseHeaders = <String>[
-        rangeStart > 0 ? 'HTTP/1.1 206 Partial Content' : 'HTTP/1.1 200 OK',
+      int requestRangeStart = int.tryParse(rangeMatch?.group(1) ?? '0') ?? 0;
+      int requestRangeEnd = int.tryParse(rangeMatch?.group(2) ?? '0') ?? -1;
+      bool partial = requestRangeStart > 0 || requestRangeEnd > 0;
+      List<String> responseHeaders = <String>[
+        partial ? 'HTTP/1.1 206 Partial Content' : 'HTTP/1.1 200 OK',
         'Accept-Ranges: bytes',
         'Content-Type: video/mp4',
         'Connection: keep-alive',
-      ].join('\r\n');
-      await socket.append(responseHeaders);
+      ];
 
-      int startRange = rangeStart - (rangeStart % Config.segmentSize);
-      int endRange = startRange + Config.segmentSize - 1;
-      DownloadTask task = DownloadTask(
-        uri: uri,
-        startRange: startRange,
-        endRange: endRange,
-      );
-      logD(
-          'Total memory size： ${await LruCacheSingleton().memoryFormatSize()}');
-      logD('Request range： ${task.startRange}-${task.endRange}');
-
-      int retry = 3;
-      while (retry > 0) {
-        Uint8List? data = await cache(task);
-        if (data != null) {
-          if (rangeStart % Config.segmentSize != 0) {
-            data = data.sublist(rangeStart % Config.segmentSize);
-          }
-          if (data.lengthInBytes > Config.segmentSize) {
-            await deleteExceedSizeFile(task);
-            retry--;
-            continue;
-          } else {
-            await socket.append(data);
-          }
-          int count = 2;
-          while (count > 0) {
-            count--;
-            task.startRange += Config.segmentSize;
-            task.endRange = task.startRange + Config.segmentSize - 1;
-            logD('Request range： ${task.startRange}-${task.endRange}');
-            data = await cache(task);
-            if (data != null) {
-              if (data.lengthInBytes > Config.segmentSize) {
-                await deleteExceedSizeFile(task);
-                retry--;
-                break;
-              } else {
-                await socket.append(data);
-              }
-            }
-          }
-          break;
-        } else {
-          concurrent(task);
-          task.priority += 10;
-          data = await download(task);
-          if (rangeStart % Config.segmentSize != 0) {
-            data = data?.sublist(rangeStart % Config.segmentSize);
-          }
-          if (data != null) {
-            if (data.lengthInBytes > Config.segmentSize) {
-              await deleteExceedSizeFile(task);
-              retry--;
-              continue;
-            } else {
-              await socket.append(data);
-            }
-          }
-          break;
-        }
+      if (Platform.isAndroid) {
+        await parseAndroid(
+          socket,
+          uri,
+          responseHeaders,
+          requestRangeStart,
+          requestRangeEnd,
+        );
+      } else {
+        await parseIOS(
+          socket,
+          uri,
+          responseHeaders,
+          requestRangeStart,
+          requestRangeEnd,
+        );
       }
       await socket.flush();
-      logD('Return request data: $uri range: $startRange-$endRange');
       return true;
     } catch (e) {
-      logE('[UrlParserMp4] ⚠ ⚠ ⚠ parse error: $e');
+      logW('[UrlParserMp4] ⚠ ⚠ ⚠ parse error: $e');
       return false;
     } finally {
       await socket.close();
       logD('Connection closed\n');
     }
+  }
+
+  Future<void> parseAndroid(
+    Socket socket,
+    Uri uri,
+    List<String> responseHeaders,
+    int requestRangeStart,
+    int requestRangeEnd,
+  ) async {
+    int contentLength = await head(uri);
+    requestRangeEnd = contentLength - 1;
+    responseHeaders.add('content-length: ${contentLength - requestRangeStart}');
+    await socket.append(responseHeaders.join('\r\n'));
+
+    bool downloading = true;
+    int startRange =
+        requestRangeStart - (requestRangeStart % Config.segmentSize);
+    int endRange = startRange + Config.segmentSize - 1;
+    int retry = 3;
+    while (downloading) {
+      DownloadTask task = DownloadTask(
+        uri: uri,
+        startRange: startRange,
+        endRange: endRange,
+      );
+      logD('Request range：${task.startRange}-${task.endRange}');
+
+      Uint8List? data = await cache(task);
+      if (data == null) {
+        concurrent(task);
+        task.priority += 10;
+        data = await download(task);
+      }
+      if (data == null) {
+        retry--;
+        if (retry == 0) {
+          downloading = false;
+          break;
+        }
+        continue;
+      }
+
+      int startIndex = 0;
+      int? endIndex;
+      if (startRange < requestRangeStart) {
+        startIndex = requestRangeStart - startRange;
+      }
+      if (endRange > requestRangeEnd) {
+        endIndex = requestRangeEnd - startRange + 1;
+      }
+      data = data.sublist(startIndex, endIndex);
+      socket.done.then((value) {
+        downloading = false;
+      }).catchError((e) {
+        downloading = false;
+      });
+      bool success = await socket.append(data);
+      if (!success) downloading = false;
+      startRange += Config.segmentSize;
+      endRange = startRange + Config.segmentSize - 1;
+      if (startRange > requestRangeEnd) {
+        downloading = false;
+      }
+    }
+  }
+
+  Future<void> parseIOS(
+    Socket socket,
+    Uri uri,
+    List<String> responseHeaders,
+    int requestRangeStart,
+    int requestRangeEnd,
+  ) async {
+    if ((requestRangeStart == 0 && requestRangeEnd == 1) ||
+        requestRangeEnd == -1) {
+      DownloadTask task = DownloadTask(uri: uri, startRange: 0, endRange: 1);
+      Uint8List? data = await cache(task);
+      int contentLength = 0;
+      if (data != null) {
+        contentLength = int.tryParse(Utf8Codec().decode(data)) ?? 0;
+      }
+      if (contentLength == 0) {
+        contentLength = await head(uri);
+        String filePath =
+            '${await FileExt.createCachePath(task.uri.generateMd5)}'
+            '/${task.saveFileName}';
+        File file = File(filePath);
+        file.writeAsString(contentLength.toString());
+        LruCacheSingleton().storagePut(file.path, file);
+      }
+      if (requestRangeStart == 0 && requestRangeEnd == 1) {
+        responseHeaders.add('content-range: bytes 0-1/$contentLength');
+        await socket.append(responseHeaders.join('\r\n'));
+        await socket.append([0]);
+        await socket.close();
+        return;
+      } else if (requestRangeEnd == -1) {
+        requestRangeEnd = contentLength - 1;
+      }
+    }
+
+    int contentLength = requestRangeEnd - requestRangeStart + 1;
+    responseHeaders.add('content-length: $contentLength');
+    await socket.append(responseHeaders.join('\r\n'));
+    logD('content-range：$requestRangeStart-$requestRangeEnd');
+    logD('content-length：$contentLength');
+
+    bool downloading = true;
+    int startRange =
+        requestRangeStart - (requestRangeStart % Config.segmentSize);
+    int endRange = startRange + Config.segmentSize - 1;
+    int retry = 3;
+    while (downloading) {
+      DownloadTask task = DownloadTask(
+        uri: uri,
+        startRange: startRange,
+        endRange: endRange,
+      );
+      logD('Request range：${task.startRange}-${task.endRange}');
+
+      Uint8List? data = await cache(task);
+      if (data == null) {
+        concurrent(task);
+        task.priority += 10;
+        data = await download(task);
+      }
+      if (data == null) {
+        retry--;
+        if (retry == 0) {
+          downloading = false;
+          break;
+        }
+        continue;
+      }
+
+      int startIndex = 0;
+      int? endIndex;
+      if (startRange < requestRangeStart) {
+        startIndex = requestRangeStart - startRange;
+      }
+      if (endRange > requestRangeEnd) {
+        endIndex = requestRangeEnd - startRange + 1;
+      }
+      data = data.sublist(startIndex, endIndex);
+      socket.done.then((value) {
+        downloading = false;
+      }).catchError((e) {
+        downloading = false;
+      });
+      bool success = await socket.append(data);
+      if (!success) downloading = false;
+      startRange += Config.segmentSize;
+      endRange = startRange + Config.segmentSize - 1;
+      if (startRange > requestRangeEnd) {
+        downloading = false;
+      }
+    }
+  }
+
+  Future<int> head(Uri uri) async {
+    HttpClient client = HttpClient();
+    HttpClientRequest request = await client.headUrl(uri);
+    HttpClientResponse response = await request.close();
+    client.close();
+    return response.contentLength;
   }
 
   /// Delete the file if it exceeds the size limit.