+++ lib/proxy/local_proxy_server.dart
@@ -79,8 +79,7 @@ class LocalProxyServer {
           return;
         }
 
-        String redirectUrl = path.replaceAll('/?url=', '');
-        Uri originUri = redirectUrl.toOriginUri();
+        Uri originUri = path.toOriginUri();
         logD('Handling Connections ========================================> \n'
             'protocol: $protocol, method: $method, path: $path \n'
             'headers: $headers \n'
@@ -89,7 +88,7 @@ class LocalProxyServer {
         await VideoCaching.parse(socket, originUri, headers);
       }
     } catch (e) {
-      logE('⚠ ⚠ ⚠ Connections exception: $e');
+      logW('⚠ ⚠ ⚠ Connections exception: $e');
     } finally {
       await socket.close(); // 确保连接关闭
     }