+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## 0.1.9
+
+- fixed: milti-resolution m3u8 parse error
+
+## 0.1.8
+
+- fixed: toLocalUrl may parse error in some cases
+- fixed: write error: PathNotFoundException: Cannot open file (#3)
+
+## 0.1.7
+
+- fixed: parse and download mp4 failed on iOS
+
+## 0.1.6
+
+- fixed: video pre-caching concurrent error handling
+- fixed: url with query string not work
+
+## 0.1.5
+
+- fixed: isolate log print failed
+- fixed: m3u8 encrypted key saved in error directory
+- fixed: preview video with error type
+- fixed: url with query string not work
+
 ## 0.1.4
 
 - fixed: m3u8 parse EXT-X-KEY failed